Numeric.Units.Dimensional.TF

Description

Please refer to the literate Haskell code for documentation of both API and implementation.

Documentation

newtype Dimensional v d a

Constructors

Dimensional a

data DUnit

data DQuantity

type Unit = Dimensional DUnit

type Quantity = Dimensional DQuantity

(*~) :: Num a => a -> Unit d a -> Quantity d a

(/~) :: Fractional a => Quantity d a -> Unit d a -> a

data Dim l m t i th n j

type DOne = Dim Zero Zero Zero Zero Zero Zero Zero

type DLength = Dim Pos1 Zero Zero Zero Zero Zero Zero

type DMass = Dim Zero Pos1 Zero Zero Zero Zero Zero

type DTime = Dim Zero Zero Pos1 Zero Zero Zero Zero

type DElectricCurrent = Dim Zero Zero Zero Pos1 Zero Zero Zero

type DThermodynamicTemperature = Dim Zero Zero Zero Zero Pos1 Zero Zero

type DAmountOfSubstance = Dim Zero Zero Zero Zero Zero Pos1 Zero

type DLuminousIntensity = Dim Zero Zero Zero Zero Zero Zero Pos1

type Dimensionless = Quantity DOne

type Length = Quantity DLength

type Mass = Quantity DMass

type Time = Quantity DTime

type ElectricCurrent = Quantity DElectricCurrent

type ThermodynamicTemperature = Quantity DThermodynamicTemperature

type AmountOfSubstance = Quantity DAmountOfSubstance

type LuminousIntensity = Quantity DLuminousIntensity

type family Mul a b

type family Div a b

type family Pow d x

type family Root d x

(*) :: Num a => Dimensional v d a -> Dimensional v d' a -> Dimensional v (Mul d d') a

(/) :: Fractional a => Dimensional v d a -> Dimensional v d' a -> Dimensional v (Div d d') a

(^) :: (Fractional a, NumType n) => Dimensional v d a -> n -> Dimensional v (Pow d n) a

(^+) :: (Num a, NumType n) => Dimensional v d a -> n -> Dimensional v (Pow d n) a

negate :: Num a => Quantity d a -> Quantity d a

(+) :: Num a => Quantity d a -> Quantity d a -> Quantity d a

(-) :: Num a => Quantity d a -> Quantity d a -> Quantity d a

abs :: Num a => Quantity d a -> Quantity d a

nroot :: (Floating a, NumType n) => n -> Dimensional v d a -> Dimensional v (Root d n) a

sqrt :: Floating a => Dimensional v d a -> Dimensional v (Root d Pos2) a

cbrt :: Floating a => Dimensional v d a -> Dimensional v (Root d Pos3) a

(^/) :: (Floating a, NumType n) => Dimensional v d a -> n -> Dimensional v (Root d n) a

(*~~) :: (Functor f, Num a) => f a -> Unit d a -> f (Quantity d a)

(/~~) :: (Functor f, Fractional a) => f (Quantity d a) -> Unit d a -> f a

sum :: forall d a. Num a => [Quantity d a] -> Quantity d a

dimensionlessLength :: Num a => [Dimensional v d a] -> Dimensionless a

exp :: Floating a => Dimensionless a -> Dimensionless a

atanh :: Floating a => Dimensionless a -> Dimensionless a

acosh :: Floating a => Dimensionless a -> Dimensionless a

asinh :: Floating a => Dimensionless a -> Dimensionless a

tanh :: Floating a => Dimensionless a -> Dimensionless a

cosh :: Floating a => Dimensionless a -> Dimensionless a

sinh :: Floating a => Dimensionless a -> Dimensionless a

atan :: Floating a => Dimensionless a -> Dimensionless a

acos :: Floating a => Dimensionless a -> Dimensionless a

asin :: Floating a => Dimensionless a -> Dimensionless a

tan :: Floating a => Dimensionless a -> Dimensionless a

cos :: Floating a => Dimensionless a -> Dimensionless a

sin :: Floating a => Dimensionless a -> Dimensionless a

log :: Floating a => Dimensionless a -> Dimensionless a

(**) :: Floating a => Dimensionless a -> Dimensionless a -> Dimensionless a

atan2 :: RealFloat a => Quantity d a -> Quantity d a -> Dimensionless a

one :: Num a => Unit DOne a

_0 :: Num a => Quantity d a

_1 :: Num a => Dimensionless a

_9 :: Num a => Dimensionless a

_8 :: Num a => Dimensionless a

_7 :: Num a => Dimensionless a

_6 :: Num a => Dimensionless a

_5 :: Num a => Dimensionless a

_4 :: Num a => Dimensionless a

_3 :: Num a => Dimensionless a

_2 :: Num a => Dimensionless a

pi :: Floating a => Dimensionless a

tau :: Floating a => Dimensionless a

dimUnit :: NumType n => String -> n -> Maybe String

prefix :: Num a => a -> Unit d a -> Unit d a